@@ -1,10 +1,10 @@
 module github.com/ecnepsnai/otto
 
-go 1.15
+go 1.16
 
 require (
 	github.com/ecnepsnai/cron v1.2.0
-	github.com/ecnepsnai/ds v1.4.0
+	github.com/ecnepsnai/ds v1.5.0
 	github.com/ecnepsnai/limits v1.0.1
 	github.com/ecnepsnai/logtic v1.4.0
 	github.com/ecnepsnai/osquery v1.0.0
